2020 실습과제

+++ src/main/java/kr/ac/jejunu/userdao/UserDao.java
@@ -3,7 +3,7 @@ package kr.ac.jejunu.userdao;
 import java.sql.*;
 
 public class UserDao {
-    public User get(Long id) throws ClassNotFoundException, SQLException {
+    public User get(Integer id) throws ClassNotFoundException, SQLException {
         Class.forName("com.mysql.cj.jdbc.Driver");
         Connection connection =
                 DriverManager.getConnection("jdbc:mysql://localhost/jeju?serverTimezone=UTC"
@@ -16,7 +16,7 @@ public class UserDao {
         resultSet.next();
 
         User user = new User();
-        user.setId(resultSet.getLong("id"));
+        user.setId(resultSet.getInt("id"));
         user.setName(resultSet.getString("name"));
         user.setPassword(resultSet.getString("password"));
